# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Bertran - Question API",
        "version": "0.0.1"
    },
    "paths": {
        "/api/login": {
            "post": {
                "summary": "Requisitar token JWT",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "description": "Requisitar token JWT",
                                    "type": "string"
                                },
                                "password": {
                                    "description": "Requisitar token JWT",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token JWT"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "Revogar token JWT",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/refresh_token": {
            "post": {
                "summary": "Renovar token JWT",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token JWT"
                    }
                }
            }
        },
        "/api/language": {
            "get": {
                "summary": "Display a listing of the resource.",
                "description": "Listar Language",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cole\u00e7\u00e3o de language"
                    }
                }
            },
            "post": {
                "summary": "Store a newly created resource in storage.",
                "description": "Criar Language",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "choes": {
                                    "description": "Store a newly created resource in storage.",
                                    "type": "string"
                                },
                                "votes": {
                                    "description": "Store a newly created resource in storage.",
                                    "type": "number"
                                },
                                "question_id": {
                                    "description": "Store a newly created resource in storage.",
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Language criada"
                    }
                }
            }
        },
        "/api/language/{id}": {
            "get": {
                "summary": "Display the specified resource.",
                "description": "Listar uma categoria",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Language encontrada"
                    }
                }
            },
            "put": {
                "summary": "Update the specified resource in storage.",
                "description": "Atualizar language",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "choes": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                },
                                "votes": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "number"
                                },
                                "question_id": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Language atualizada"
                    }
                }
            },
            "delete": {
                "summary": "Remove the specified resource from storage.",
                "description": "Excluir language",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/question": {
            "get": {
                "summary": "Display a listing of the resource.",
                "description": "Listar question",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cole\u00e7\u00e3o de question"
                    }
                }
            },
            "post": {
                "summary": "Store a newly created resource in storage.",
                "description": "Criar Question",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Store a newly created resource in storage.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Question criada"
                    }
                }
            }
        },
        "/api/question/{id}": {
            "get": {
                "summary": "Display the specified resource.",
                "description": "Listar uma question",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "question encontrada"
                    }
                }
            },
            "put": {
                "summary": "Update the specified resource in storage.",
                "description": "Atualizar Question",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Update the specified resource in storage.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Quest\u00f5es atualizada"
                    }
                }
            },
            "delete": {
                "summary": "Remove the specified resource from storage.",
                "description": "Excluir question",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        }
    },
    "definitions": {}
}